feat: Expose ClientTrustConfig as a public class
