+++ sigstore/_internal/rekor/client_v2.py
@@ -21,6 +21,7 @@ from __future__ import annotations
 import base64
 import json
 import logging
+import typing
 
 import requests
 from cryptography.hazmat.primitives import serialization
@@ -38,7 +39,9 @@ from sigstore._internal.rekor import (
 )
 from sigstore.dsse import Envelope
 from sigstore.hashes import Hashed
-from sigstore.models import TransparencyLogEntry
+
+if typing.TYPE_CHECKING:
+    from sigstore.models import TransparencyLogEntry
 
 _logger = logging.getLogger(__name__)
