+++ test/unit/test_sign.py
@@ -21,10 +21,10 @@ from sigstore_models.common.v1 import HashAlgorithm
 
 import sigstore.oidc
 from sigstore._internal.timestamp import TimestampAuthorityClient
-from sigstore._internal.trust import ClientTrustConfig
 from sigstore.dsse import StatementBuilder, Subject
 from sigstore.errors import VerificationError
 from sigstore.hashes import Hashed
+from sigstore.models import ClientTrustConfig
 from sigstore.sign import SigningContext
 from sigstore.verify.policy import UnsafeNoOp
